Fixed the flaky breadcrumbs tests
