+++ test/functional/cypress/support/assertions.js
@@ -23,19 +23,46 @@ const assertValueTable = (dataAutoId, expected) => {
   })
 }
 
-const assertBreadcrumbs = (expected) => {
-  Object.keys(expected).map((text, index) => {
-    const href = expected[text]
-    const breadcrumbNumber = index + 1
-
-    if (href) {
-      cy.get(selectors.breadcrumbs.item.byNumber(breadcrumbNumber)).should('have.text', text)
-      cy.get(selectors.breadcrumbs.item.byNumber(breadcrumbNumber)).should('have.attr', 'href', href)
-    } else {
-      cy.get(selectors.breadcrumbs.item.last()).should('have.text', text)
-    }
-  })
+/**
+ * @description Asserts the presence of breadcrumbs with minimal knowledge about
+ * implementation details e.g. class names and ids.
+ * @param {Object} specs - A map of expected breadcrumb item labels to hrefs.
+ * @example
+ * // Asserts that breadcrumbs: Home > Foo > Bar > Baz exist and that they have
+ * // the expected texts and hrefs.
+ * it('Should render foo > bar > baz breadcrumbs' =>
+ *   assertBreadcrumbs({
+ *     'Home': '/',
+ *     'Foo': '/foo',
+ *     'Bar': '/bar',
+ *     'Baz': undefined,
+ *   })
+ * )
+ */
+const assertBreadcrumbs = specs => {
+  const entries = Object.entries(specs)
+  cy
+    .contains(Object.keys(specs).join(''))
+    .children('li')
+    .should('have.length', entries.length)
+    .each((x, i) => {
+      const [expectedText, expectedHref] = entries[i]
+      cy
+        .get(x)
+        .contains(expectedText)
+        .invoke('attr', 'href')
+        .should('eq', expectedHref || undefined)
+    })
 }
+
+/**
+ * @description Same as asserBreadcrumbs but already wrapped in an `it` block.
+ * @param {Object} specs - A map of expected breadcrumb item labels to hrefs.
+ */
+const testBreadcrumbs = specs =>
+  it('Should render breadcrumbs', () =>
+    assertBreadcrumbs(specs)
+  )
 
 const assertFieldInput = ({ name, label = null, value = null }) => {
   const inputElement = cy.get(`input[name="${name}"]`)
@@ -79,6 +106,7 @@ module.exports = {
   assertKeyValueTable,
   assertValueTable,
   assertBreadcrumbs,
+  testBreadcrumbs,
   assertFieldInput,
   assertFieldUneditable,
   assertLocalHeader,
